Revert "Next audio step."
This reverts commit 79e8f49bb82b6729b55387bdd815a3c4889840dc.

Conflicts:
	sound/soc/codecs/tlv320aic31xx.c

+++ sound/soc/codecs/aic3xxx/aic3xxx_cfw_ops.c
@@ -1124,7 +1124,7 @@ static int aic3xxx_cfw_release(struct inode *in, struct file *filp)
 	ps->is_open--;
 	return ps->is_open;
 }
-static long aic3xxx_cfw_ioctl(struct file *filp,
+static int aic3xxx_cfw_ioctl(struct file *filp,
 			unsigned int cmd, unsigned long arg)
 {
 	return 0;
